Handle patches manually in pmm-install-python
We need to provide a custom patch for the `_decimal` module. However,
if we use the switch `--patch` from `pyenv install` we lose the builtin
patches that are applied when no custom patch is passed.

For this reason, this commit creates a custom patch file that consists
of the pyenv builtin patches plus the `_decimal` patch when required.